Avoid a warning on Visual Studio.
SVN-Revision: 4082

+++ libarchive/test/read_open_memory.c
@@ -145,7 +145,7 @@ memory_read(struct archive *a, void *client_data, const void **buff)
 		buff = NULL;
 		return 0;
 	}
-	if (size > mine->read_size)
+	if ((size_t)size > mine->read_size)
 		size = mine->read_size;
 	else
 		memset(mine->copy_buff, 0xA5, mine->copy_buff_size);
